+++ flipper/flytrap/scenes/flytrap_scene_config.h
@@ -1,3 +1,4 @@
 ADD_SCENE(flytrap, main_menu, MainMenu)
 ADD_SCENE(flytrap, ssid_input, SsidInput)
 ADD_SCENE(flytrap, live, Live)
+ADD_SCENE(flytrap, textview, TextView)